refactor interactions and utils
